requirements.txt
asgiref
Django
Pillow
psycopg2-binary
sqlparse
django_crontab
tzdata
python-dotenv
djangorestframework
djangorestframework-simplejwt
django-filter
drf_yasg
pyyaml
django-cors-headers
requests
celery
redis
django_celery_beat
python-telegram-bot
telegram
coverage
telebot